+++ src/components/BulkEditList/BulkEditListFilters/BulkEditListFilters.js
@@ -53,8 +53,8 @@ export const BulkEditListFilters = ({
     hasCsvViewPerms,
     hasInAppUsersEditPerms,
     hasInAppViewPerms,
-    hasViewInAppPerms,
     hasOnlyViewCsvPerms,
+    hasAnyInventoryWithInAppView,
   } = permissions;
   const showCallout = useShowCallout();
   const history = useHistory();
@@ -235,7 +235,10 @@ export const BulkEditListFilters = ({
   const renderListSelect = () => (
     <ListSelect
       value={recordIdentifier}
-      disabled={getIsDisabledByPerm(capabilities, isSelectIdentifiersDisabled, hasOnlyViewCsvPerms, hasViewInAppPerms)}
+      disabled={getIsDisabledByPerm(capabilities,
+        isSelectIdentifiersDisabled,
+        hasOnlyViewCsvPerms,
+        hasAnyInventoryWithInAppView)}
       onChange={handleRecordIdentifierChange}
       capabilities={capabilities}
     />
@@ -249,7 +252,7 @@ export const BulkEditListFilters = ({
       handleDrop={handleDrop}
       isDropZoneDisabled={isDropZoneDisabled
           ||
-          getIsDisabledByPerm(capabilities, isDropZoneDisabledPerm, hasOnlyViewCsvPerms, hasViewInAppPerms)}
+          getIsDisabledByPerm(capabilities, isDropZoneDisabledPerm, hasOnlyViewCsvPerms, hasAnyInventoryWithInAppView)}
       recordIdentifier={recordIdentifier}
       handleDragLeave={handleDragLeave}
       handleDragEnter={handleDragEnter}